% Auto-generated: do not edit by hand
\name{dccPagerPager}

\alias{dccPagerPager}

\title{Pager component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dccPagerPager(id=NULL, maxValue=NULL, minValue=NULL, symbols=NULL,
value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{maxValue}{Numeric. A label that will be printed when this component is rendered.}

\item{minValue}{Numeric. The minimum number of elements displayed in the input.}

\item{symbols}{List of characters. The text symbols for the buttons.}

\item{value}{List of numerics. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

